Significant changes possibly breaking behavior.
using an overrides.txt file which is comprised of my own overrides appended to the vanillapack overrides file and named overrides.txt, along with some png files in the textures folder, a few are from vanillapack and the others are for the lode rake and star flower adjustments.  The 'pfy' and 'tri' options still exist, those would need to be renamed differently now. The readme is not accurate right now.